requirements.txt
PANDAS==1.3.4
MATPLOTLIB==3.5.1
NUMPY==1.26.4
PSYCOPG2-BINARY==2.9.1
SEABORN==0.13.2
SQLALCHEMY>=1.4.28